Add TRR test trajectories and rename test file
